feat(query-builder): Allow pasting without requiring user permissions
